# api/openapi.yaml
openapi: 3.1.0
info:
  description: REST and Server-Sent Events API of the Sift server. Errors are RFC
    9457 problem details (`application/problem+json`).
  title: Sift Server API
  version: v1
security:
- bearerAuth: []
...
paths:
  /api/v1/agents:
    get:
      operationId: agentRunList
      parameters:
      - in: query
        name: kind
        required: false
        schema:
          type: string
          enum:
          - CODE_REVIEW
      - in: query
        name: phase
        required: false
        schema:
          type: string
          enum:
          - CREATED
          - PENDING
          - RUNNING
          - SUCCESS
          - FAILED
          - CANCELLED
      - in: query
        name: repositoryId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: mine
        required: false
        schema:
          type: boolean
          default: false
      - in: query
        name: createdBy
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseAgentRunResponse"
          description: "One page of runs, newest first"
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: '`mine` and `createdBy` name different users'
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
      summary: List agent runs
      tags:
      - agents
    post:
      operationId: agentRunCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRunRequest"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
          description: Run accepted; `Location` points at the new run
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Validation failed
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Repository not found
      summary: Start an agent run
      tags:
      - agents
  /api/v1/agents/watch:
    get:
      description: "Open-ended `text/event-stream`. Each event is named after `AgentRunEvent.type`\
        \ (`SNAPSHOT` or `UPDATED`), carries an `AgentRunEvent` JSON payload as `data`\
        \ and the run's `updatedAt` epoch millis as `id`; send it back as `Last-Event-ID`\
        \ to resume. `:heartbeat` comment frames keep the stream alive."
      operationId: agentWatchWatch
      parameters:
      - in: query
        name: agentId
        required: false
        schema:
          type: string
          format: uuid
      - in: query
        name: kind
        required: false
        schema:
          type: string
          enum:
          - CODE_REVIEW
      - in: query
        name: mine
        required: false
        schema:
          type: boolean
          default: false
      - in: query
        name: createdBy
        required: false
        schema:
          type: string
          format: uuid
      - in: header
        name: Last-Event-ID
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/AgentRunEvent"
          description: Event stream; each `data` frame is one `AgentRunEvent`
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: '`mine` and `createdBy` name different users'
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
      summary: Watch agent runs (Server-Sent Events)
      tags:
      - agents
  /api/v1/agents/{id}:
    delete:
      description: A run that is still active is cancelled first; its review result
        and findings are deleted too.
      operationId: agentRunDelete
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Run deleted
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Run not found
      summary: Delete an agent run
      tags:
      - agents
    get:
      operationId: agentRunGet
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
          description: The run
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Run not found
      summary: Get an agent run
      tags:
      - agents
  /api/v1/agents/{id}/cancel:
    post:
      operationId: agentRunCancel
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResponse"
          description: Cancellation requested
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Missing or invalid bearer token
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Run not found
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
          description: Run already reached a terminal phase
      summary: Cancel an agent run
      tags:
      - agents
...
components:
  schemas:
    AgentRunEvent:
      type: object
      properties:
        run:
          $ref: "#/components/schemas/AgentRunResponse"
        type:
          type: string
          enum:
          - SNAPSHOT
          - UPDATED
    AgentRunResponse:
      type: object
      properties:
        completedAt:
          type:
          - string
          - "null"
          format: date-time
        crName:
          type: string
        crUid:
          type:
          - string
          - "null"
        createdAt:
          type: string
          format: date-time
        createdBy:
          oneOf:
          - $ref: "#/components/schemas/RunCreatorResponse"
          - type: "null"
        executionId:
          type:
          - string
          - "null"
        generation:
          type:
          - integer
          - "null"
          format: int64
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
          - CODE_REVIEW
        message:
          type:
          - string
          - "null"
        phase:
          type: string
          enum:
          - CREATED
          - PENDING
          - RUNNING
          - SUCCESS
          - FAILED
          - CANCELLED
        reason:
          type:
          - string
          - "null"
        repositoryId:
          type:
          - string
          - "null"
          format: uuid
        source:
          type: string
          enum:
          - API
          - EXTERNAL
        spec:
          type: object
          additionalProperties: {}
          description: "Kind-specific run parameters, e.g. `CodeReviewRunSpec` for\
            \ `CODE_REVIEW`"
        startedAt:
          type:
          - string
          - "null"
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateAgentRunRequest:
      type: object
      properties:
        baseBranch:
          type: string
          minLength: 1
        branch:
          type: string
          minLength: 1
        commitSha:
          type: string
          pattern: "^[0-9a-fA-F]{40}$"
        kind:
          type: string
          enum:
          - CODE_REVIEW
        pullRequest:
          type:
          - string
          - "null"
        repositoryId:
          type: string
          format: uuid
      required:
      - baseBranch
      - branch
...
    PageResponseAgentRunResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AgentRunResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
...
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...
    RunCreatorResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Access token issued by the OIDC provider announced by `GET /api/v1/auth/config`.
      scheme: bearer
      type: http